test(sync): consolidate squash-merge tests into sync_test.go
The pressure tests for squash merges live alongside the existing
sync tests rather than in a separate file. Helpers (markPrMerged,
disableCommitSigning) move with them.
